refactor: move config and related files to relevant directories
