+++ src/components/Configs/ConfigComponents/index.tsx
@@ -2,7 +2,7 @@ import { FaExclamationTriangle } from "react-icons/fa";
 import { Link } from "react-router-dom";
 import { useConfigComponentsRelationshipQuery } from "../../../api/query-hooks/useConfigComponentsRelationshipQuery";
 import { CountBadge } from "../../../ui/Badge/CountBadge";
-import CollapsiblePanel from "../../CollapsiblePanel";
+import CollapsiblePanel from "../../../ui/CollapsiblePanel/CollapsiblePanel";
 import { Icon } from "../../Icon";
 import { TopologyIcon } from "../../Icons/TopologyIcon";
 import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
