+++ src/components/Configs/Sidebar/ConfigChanges/index.tsx
@@ -3,8 +3,8 @@ import { useEffect, useMemo, useState } from "react";
 import { useGetConfigChangesByConfigIdQuery } from "../../../../api/query-hooks";
 import { ConfigChange } from "../../../../api/types/configs";
 import PillBadge from "../../../../ui/Badge/PillBadge";
+import CollapsiblePanel from "../../../../ui/CollapsiblePanel/CollapsiblePanel";
 import { CreatedAtCell } from "../../../../ui/table/DateCells";
-import CollapsiblePanel from "../../../CollapsiblePanel";
 import EmptyState from "../../../EmptyState";
 import { Icon } from "../../../Icon";
 import { InfiniteTable } from "../../../InfiniteTable/InfiniteTable";
