+++ src/components/Configs/Sidebar/ConfigInsights/index.tsx
@@ -1,9 +1,9 @@
 import { MdOutlineInsights } from "react-icons/md";
 import { useGetConfigInsights } from "../../../../api/query-hooks";
 import PillBadge from "../../../../ui/Badge/PillBadge";
-import CollapsiblePanel from "../../../CollapsiblePanel";
-import InsightsDetails from "../../../Insights/Insights";
+import CollapsiblePanel from "../../../../ui/CollapsiblePanel/CollapsiblePanel";
 import Title from "../../../Title/title";
+import InsightsDetails from "../Insights/Insights";
 
 type Props = {
   configID: string;
