+++ src/components/IncidentDetails/DefinitionOfDone/IncidentsDefinitionOfDone.tsx
@@ -8,8 +8,8 @@ import { Evidence } from "../../../api/types/evidence";
 import { useIncidentState } from "../../../store/incident.state";
 import { CountBadge } from "../../../ui/Badge/CountBadge";
 import { ClickableSvg } from "../../../ui/ClickableSvg/ClickableSvg";
+import CollapsiblePanel from "../../../ui/CollapsiblePanel/CollapsiblePanel";
 import { dateSortHelper } from "../../../utils/date";
-import CollapsiblePanel from "../../CollapsiblePanel";
 import { ConfirmationPromptDialog } from "../../Dialogs/ConfirmationPromptDialog";
 import EmptyState from "../../EmptyState";
 import Title from "../../Title/title";
